+++ cli/src/index.ts
@@ -17,7 +17,7 @@ import { AstarAPI } from "./lib/api";
 import { c } from "./lib/ui";
 import { resolve } from "path";
 
-export const VERSION = "0.0.17";
+export const VERSION = "0.0.18";
 
 async function showDashboard() {
   const status = await getAuthStatus();